Replaced int codes with enums for movement recording codes

+++ includes/correlations3Axes.h
@@ -12,7 +12,9 @@
 #ifndef CORRELATION_3_AXES_H
 #define CORRELATION_3_AXES_H
 
+#include <stdint.h>
 #include "signaux.h"
+
 
 void correler3AxesDemo(Signal3AxesPtr* p_signalACorrelerPtr, Signal3AxesReference* p_signal3AxesReference);
 
@@ -20,7 +22,7 @@ void correler3AxesDemo(Signal3AxesPtr* p_signalACorrelerPtr, Signal3AxesReferenc
  * Effectue la correlation sur les 3 axes et retourne 1 si le signal à correler est suffisament semblable
  * au signal de référence
  */
-int correler3Axes(Signal3AxesPtr* p_signalACorrelerPtr, Signal3AxesReference* p_signal3AxesReference);
+uint8_t correler3Axes(Signal3AxesPtr* p_signalACorrelerPtr, Signal3AxesReference* p_signal3AxesReference);
 
 /*
  * Remplit la structure Signal3AxesReference à partir des données comprises dans ses champs x,y et z
